Implement display site links
