+++ store/siteLinks.js
@@ -1,0 +1,9 @@
+export const state = () => ({
+  siteLinks: []
+})
+
+export const mutations = {
+  set(state, siteLinks) {
+    state.siteLinks = siteLinks || []
+  }
+}
